// max-most-frontend/src/Interfaces/Company.ts
export interface ProductTransaction {
  running_total: string;
  ordered: string;
  name: string;
  number: string;
  batch_number: string;
  expiry_date: Date | string;
  quantity: number;
  type_t: string;
  is_adjustment: boolean;
}